#443: Moved the PassiveDualWieldWeaponCategoryUnlockParser into the  model folder
 -
